add q33
Actually it doesn't have to be so brute force, but anyway since there is
not much calculation.

If the numerator is ab, and the denominator is cd, it can be proved that
it must be b==c or a==d; the other two options are not legal (a==c or
b==d).